refactor: add usePlayerName hook
