+++ packages/app/src/components/NoteCard.tsx
@@ -1,19 +1,17 @@
 import { encodeBlock } from "@dust/world/internal";
-import { getRecord } from "@latticexyz/stash/internal";
-import { useMemo } from "react";
 import { toast } from "sonner";
-import { hexToString, zeroAddress } from "viem";
 
 import { useCopy } from "@/common/useCopy";
 import { useDustClient } from "@/common/useDustClient";
+import { usePlayerName } from "@/common/usePlayerName";
 import { Badge } from "@/components/ui/badge";
 import { cn } from "@/lib/utils";
-import { stash, tables } from "@/mud/stash";
 import { formatDate, shortenAddress } from "@/utils/helpers";
 import type { Post } from "@/utils/types";
 
 export const NoteCard = ({ note }: { note: Post }) => {
   const { data: dustClient } = useDustClient();
+  const { playerName } = usePlayerName(note.owner);
   const { copyToClipboard } = useCopy();
 
   // Set waypoint for an note by encoding its block coords into an EntityId
@@ -46,19 +44,6 @@ export const NoteCard = ({ note }: { note: Post }) => {
     }
   };
 
-  const author = useMemo(() => {
-    const ownerUsername = getRecord({
-      stash,
-      table: tables.PlayerName,
-      key: { player: (note?.owner ?? zeroAddress) as `0x${string}` },
-    })?.name;
-
-    if (ownerUsername) {
-      return hexToString(ownerUsername).replace(/\0+$/, "");
-    }
-    return "Anonymous";
-  }, [note?.owner]);
-
   return (
     <div
       key={note.id}
@@ -88,7 +73,7 @@ export const NoteCard = ({ note }: { note: Post }) => {
               toast.success(`Copied ${shortenAddress(note.owner)}`);
             }}
           >
-            @{author}
+            @{playerName}
           </button>
         </div>
       </div>
