+++ packages/app/src/components/editor/ArticleWizardStep2.tsx
@@ -1,8 +1,4 @@
-import { getRecord } from "@latticexyz/stash/internal";
-import { useMemo } from "react";
-import { hexToString, zeroAddress } from "viem";
-
-import { stash, tables } from "@/mud/stash";
+import { usePlayerName } from "@/common/usePlayerName";
 import { uriToHttp } from "@/utils/helpers";
 
 interface Props {
@@ -24,18 +20,7 @@ export default function ArticleWizardStep2({
   authorAddress,
   category,
 }: Props) {
-  const author = useMemo(() => {
-    const ownerUsername = getRecord({
-      stash,
-      table: tables.PlayerName,
-      key: { player: (authorAddress ?? zeroAddress) as `0x${string}` },
-    })?.name;
-
-    if (ownerUsername) {
-      return hexToString(ownerUsername).replace(/\0+$/, "");
-    }
-    return "Anonymous";
-  }, [authorAddress]);
+  const { playerName } = usePlayerName(authorAddress);
 
   return (
     <div className="max-h-[60vh] overflow-y-auto bg-panel border border-neutral-200 rounded p-4">
@@ -50,7 +35,7 @@ export default function ArticleWizardStep2({
         <h1 className={"font-heading text-2xl leading-tight"}>
           {title || "Untitled"}
         </h1>
-        <div className="font-accent text-[10px] text-neutral-700">{`By ${author}`}</div>
+        <div className="font-accent text-[10px] text-neutral-700">{`By ${playerName}`}</div>
         {category && (
           <div className="mt-1">
             <span className="font-accent bg-neutral-100 border border-neutral-900 px-2 py-1 rounded-[3px] text-[10px] tracking-wider uppercase">
